Moved created files into KelsonD directory
