+++ clients/reference/requirements.txt
@@ -1,0 +1,2 @@
+grpcio==1.33.2
+protobuf==3.14.0
